refactor: move listeners to the main library
